opensearch-elasticsearch/requirements.txt
apscheduler
asteval
boto3
elasticsearch
opensearch-py
opensearch-dsl
flask
flask_autoindex
flask-apscheduler
flask-cors
flask-httpauth
flask-wtf
gunicorn
numpy==1.22.0
pandas
pytz
pyyaml
requests
atpbar
tqdm
enlighten
